Feature/dpo 2114 spring boot (#90)
* DPO-2114 update dependencies, javax -> jakarta changes

* DPO-2114 change to jakarta namespaces

* DPO-2114 fix version

* DPO-2114 fix namespaces and versions

* DPO-2114 add javax validation

* DPO-2114 validation-api version

* DPO-2114 fixes

* DPO-2114 remove @ConstructorBinding

* DPO-2114 implement missing methods

* DPO-2114 new openapi, new spatial, new hibernate-types, fix sequence usage

* DPO-2114 disable blackbird, use casting

* DPO-2114 native query returns now as instant

* DPO-2114 native query returns Long

* DPO-2114 wiremock-standalone, add cast

* DPO-2114 some casting, fix tests, use java 17 on install

* DPO-2114 remove commented

* DPO-2114 update ubuntu version

* DPO-2114 fix build command

* DPO-2114 fixes

+++ src/test/java/fi/livi/digitraffic/tie/service/v3/maintenance/V3MaintenanceTrackingServiceTestHelper.java
@@ -21,7 +21,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import javax.persistence.EntityManager;
+import jakarta.persistence.EntityManager;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.tuple.Pair;
@@ -556,12 +556,12 @@ public class V3MaintenanceTrackingServiceTestHelper {
                     "ST_GeometryFromText('LINESTRING(20.10 64.10 0, 20.0 64.0 0)', " + GeometryConstants.SRID + "), " +
                     workMachineId + ", 'dummy', now(), now(), now(), true)" )
             .executeUpdate();
-        final long id = ((BigInteger) entityManager.createNativeQuery(
+        final Long id = (Long)entityManager.createNativeQuery(
             "select id " +
                 "from road.public.maintenance_tracking " +
                 "where domain = '" + domain + "' " +
                 "order by id desc " +
-                "limit 1").getSingleResult()).longValue();
+                "limit 1").getSingleResult();
         entityManager.createNativeQuery(
                 "INSERT INTO road.public.maintenance_tracking_task(maintenance_tracking_id, task)\n" +
                     "VALUES (" + id + ", '" + BRUSHING.name() + "')")
